Moved Login function to a separated file
