+++ sdk/requests/index.js
@@ -1,8 +1,8 @@
-import { Linking } from 'react-native';
 import { encode } from 'base-64';
 import { fetch } from 'react-native-ssl-pinning';
 import { getParameters, setParameters } from '../configuration';
-import { loginEndpoint, tokenEndpoint } from './endpoints';
+import { tokenEndpoint } from './endpoints';
+import login from './login';
 
 export const REQUEST_TYPES = {
   LOGIN: 'login',
@@ -13,50 +13,7 @@ const makeRequest = async type => {
   const parameters = getParameters();
   switch (type) {
     case REQUEST_TYPES.LOGIN: {
-      let resolveFunction;
-      let rejectFunction;
-
-      // Se crea una promise para devolver, y se guardan
-      // referencias a sus funciones de reject y resolve.
-      const promise = new Promise((resolve, reject) => {
-        resolveFunction = resolve;
-        rejectFunction = reject;
-      });
-
-      // Handler para el evento url.
-      const handleOpenUrl = event => {
-        // Obtiene el auth code a partir de la url a la que
-        // redirige el browser luego de realizado el login.
-        const code = event.url.match(/\?code=([^&]+)/);
-
-        // Si existe el código, se guarda y se resuelve la promise
-        // si no, se rechaza la promise con un error.
-        if (code) {
-          setParameters({ code: code[1] });
-          resolveFunction(code[1]);
-        } else rejectFunction(Error('Invalid authorization code'));
-
-        // Se elimina el handler para los eventos url.
-        Linking.removeEventListener('url', handleOpenUrl);
-      };
-
-      try {
-        // Se agrega el handler para eventos url.
-        Linking.addEventListener('url', handleOpenUrl);
-        // Si hay un clientId setteado, se abre el browser
-        // para realizar la autenticación con idUruguay.
-        if (parameters.clientId) await Linking.openURL(loginEndpoint());
-        else {
-          // En caso de error, se elimina el handler y rechaza la promise.
-          Linking.removeEventListener('url', handleOpenUrl);
-          rejectFunction(Error("Couldn't make request"));
-        }
-      } catch (error) {
-        // En caso de error, se elimina el handler y rechaza la promise.
-        Linking.removeEventListener('url', handleOpenUrl);
-        rejectFunction(Error("Couldn't make request"));
-      }
-      return promise;
+      return login();
     }
     case REQUEST_TYPES.GET_TOKEN: {
       // Codificar en base64 el clientId y el clientSecret,
